+++ app/src/main/AndroidManifest.xml
@@ -18,16 +18,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".Activities.ExerciseDetailActivity"
-            android:label="@string/title_activity_exercise_detail"
-            android:theme="@style/AppTheme.NoActionBar" >
-        </activity>
-        <activity
-            android:name=".Activities.WorkoutDetailActivity"
-            android:label="@string/title_activity_workout_detail"
-            android:theme="@style/AppTheme.NoActionBar" >
-        </activity>
     </application>
 
 </manifest>
