+++ app/src/main/res/layout-land/activity_workouts.xml
@@ -7,7 +7,7 @@
 
 <fragment
     android:id="@+id/list"
-    android:name="edu.uco.schambers4.octane.Fragments.WorkoutDetailFragment"
+    android:name="edu.uco.schambers4.octane.Fragments.Workouts.WorkoutDetailFragment"
     android:layout_width="0dp"
     android:layout_height="match_parent"
     android:layout_weight="1"
