+++ app/src/main/res/layout/activity_exercise_detail.xml
@@ -6,8 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="edu.uco.schambers4.octane.Activities.ExerciseDetailActivity">
-
+    >
     <android.support.design.widget.AppBarLayout
         android:layout_height="wrap_content"
         android:layout_width="match_parent"
