+++ app/src/main/res/layout/content_exercise_detail.xml
@@ -11,6 +11,6 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:showIn="@layout/activity_exercise_detail"
-    tools:context="edu.uco.schambers4.octane.Activities.ExerciseDetailActivity">
+    >
 
 </RelativeLayout>
