+++ app/src/main/res/layout/fragment_exercise_detail.xml
@@ -2,7 +2,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="edu.uco.schambers4.octane.Fragments.ExerciseDetailFragment">
+             tools:context="edu.uco.schambers4.octane.Fragments.Exercises.ExerciseDetailFragment">
 
 <TableRow
     android:background="@color/material_blue_grey_800">
