+++ app/src/main/res/layout/fragment_workout_detail.xml
@@ -2,7 +2,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="edu.uco.schambers4.octane.Fragments.WorkoutDetailFragment">
+             tools:context="edu.uco.schambers4.octane.Fragments.Workouts.WorkoutDetailFragment">
 
 <TableRow
     android:background="@color/material_blue_grey_800">
